showtime!

+++ Web/Controllers/RecipeController.cs
@@ -55,9 +55,9 @@ namespace Web.Controllers
             _reportService = reportService;
         }
 
-        // CrUD
+        // Create / Update / Delete Recipe
         [HttpGet]
-        [Authorize]
+        [Authorize(Roles = "Admin, Manager, User")]
         public IActionResult AddorEdit(long recipeId = 0)
         {
             if (recipeId == 0)
@@ -175,10 +175,11 @@ namespace Web.Controllers
             string wwwRootPath = _webHostEnvironment.WebRootPath;
             string imagePath = _recipeService.GetRecipe(_postService.GetPost(postId).RecipeId).ImagePath;
             _postService.DeletePost(postId);
-            if (imagePath != "/Image/test1.jpg" && imagePath != "/Image/emptyImage.jpg") { System.IO.File.Delete(wwwRootPath + imagePath);}
+            if (imagePath != "/Image/test1.jpg" && imagePath != "/Image/emptyImage.png") { System.IO.File.Delete(wwwRootPath + imagePath);}
             return Ok();
         }
-        // Partial
+
+        // Partial work
         [HttpPost]
         public async Task<IActionResult> PartialPost(string typePar, string inpPar, int catPar, int difPar)
         {
@@ -300,6 +301,7 @@ namespace Web.Controllers
             return PartialView("_ShowObject", model);
 
         }
+
         // Work with reports
         [Authorize(Roles = "User")]
         [HttpPost]
@@ -315,10 +317,9 @@ namespace Web.Controllers
             var userId = HttpContext.User.Claims.FirstOrDefault(a => a.Type == ClaimTypes.NameIdentifier).Value;
             return Ok(_reportService.ReportPost(postId, userId, target.Id));
         }
-        // =================
 
 
-        // 
+        // Change Language and subscribtion work
         [HttpGet]
         public IActionResult SetLanguage(string culture, string returnAction)
         {
@@ -329,23 +330,6 @@ namespace Web.Controllers
             );
             return Redirect(returnAction);
         }
-
-        [Route("")]
-        [Route("Recipe/Show")]
-        [HttpGet]
-        public IActionResult Show()
-        {
-            ViewData["returnAction"] = "/Recipe/Show";
-            return View();
-        }
-
-        
-
-
-
-
-
-
         [HttpPost]
         [Authorize]
         public IActionResult Subscribe(long postId, string subFlag)
@@ -355,19 +339,18 @@ namespace Web.Controllers
             else _postService.UnsubscribePost(postId, userId);
             return Ok();
         }
-        
 
-
+        // Insert Tips, Methods, Ingredients, Image into db
         private string GetImagePath(IFormFile recipeImage, string prevImagePath = null)
         {
             if (recipeImage != null)
             {
                 string wwwRootPath = _webHostEnvironment.WebRootPath;
-                if (prevImagePath != null)
+                if (prevImagePath != null && prevImagePath != "emptyImage.png")
                 {
                     System.IO.File.Delete(Path.Combine(wwwRootPath + "/Image/", prevImagePath));
                 }
-                
+
                 string fileName = Path.GetFileNameWithoutExtension(recipeImage.FileName);
                 string extension = Path.GetExtension(recipeImage.FileName);
                 fileName = fileName + DateTime.Now.ToString("eemmssfff") + extension;
@@ -417,70 +400,13 @@ namespace Web.Controllers
             }
         }
 
-        [Authorize]
-        public IActionResult AddConst()
+        [Route("")]
+        [Route("Recipe/Show")]
+        [HttpGet]
+        public IActionResult Show()
         {
-            var userId = HttpContext.User.Claims.FirstOrDefault(a => a.Type == ClaimTypes.NameIdentifier).Value;
-            Recipe recipeEntity = new Recipe()
-            {
-                Title = "Плов со свининой",
-                Description = "Плов - это простая еда. И готовится плов легко и просто! Без всяких понтов и заморочек!",
-                Category = 3,
-                PrepTime = 20,
-                CookTime = 80,
-                Marinade = 0,
-                Difficulty = 2,
-                ImagePath = "/Image/test1.jpg",
-                UserId = userId
-            };
-            _recipeService.InsertRecipe(recipeEntity);
-
-            var ingredient1 = new Ingredient() { Name = "Морковь 2 штуки", Recipe = recipeEntity };
-            var ingredient2 = new Ingredient() { Name = "Свинина 500 г", Recipe = recipeEntity };
-            var ingredient3 = new Ingredient() { Name = "Репчатый лук 2 штуки", Recipe = recipeEntity };
-            var ingredient4 = new Ingredient() { Name = "Рис 2 стакана", Recipe = recipeEntity };
-            var ingredient5 = new Ingredient() { Name = "Растительное масло 50 мл", Recipe = recipeEntity };
-            var ingredient6 = new Ingredient() { Name = "Соль по вкусу", Recipe = recipeEntity };
-            var ingredient7 = new Ingredient() { Name = "Специи для плова", Recipe = recipeEntity };
-            var ingredient8 = new Ingredient() { Name = "Чеснок 1 головка", Recipe = recipeEntity };
-            var method1 = new Method() { Name = "Мясо вымыть, обсушить, разрезать на кусочки.", Recipe = recipeEntity };
-            var method2 = new Method() { Name = "Очистить, помыть и нарезать тонкими полукольцами репчатый лук.", Recipe = recipeEntity };
-            var method3 = new Method() { Name = "Морковь очистить, помыть и нарезать соломкой.", Recipe = recipeEntity };
-            var method4 = new Method() { Name = "Сделать зирвак (это основа плова, а именно - мясо, морковь, лук и специи). Для этого разогреть казанок. Налить растительное масло. Хорошо его прожарить. Выложить подготовленный лук. Жарить до золотистого цвета около 5–7 минут, помешивая", Recipe = recipeEntity };
-            var method5 = new Method() { Name = "Выложить подготовленное мясо. Все готовить до состояния, когда мясо покроется зажаренной корочкой, около 10 минут.", Recipe = recipeEntity };
-            var method6 = new Method() { Name = "Затем добавить морковь. Обжарить все вместе, помешивая, 3–5 минут.", Recipe = recipeEntity };
-            var method7 = new Method() { Name = "Вскипятить чайник. Обжаренные овощи и мясо залить кипятком. Посолить, поперчить, добавить специи для плова. Варить на среднем огне 20 минут.", Recipe = recipeEntity };
-            var method8 = new Method() { Name = "Положить рис, осторожно разровнять его по поверхности. В серединку поместить головку чеснока. Долить воды так, чтобы она покрывала рис выше на 2 см. Варить плов на максимальном огне без крышки почти до полного испарения жидкости, около 10–15 минут.", Recipe = recipeEntity };
-            var method9 = new Method() { Name = "Как только вода испарилась, сделать минимальный огонь. В плове сделать несколько отверстий ручкой ложки. Накрыть крышкой и оставить плов со свининой упариваться на 15 минут.", Recipe = recipeEntity };
-            var tip1 = new Tip() { Name = "Вода должна покрывать мясо с овощами где-то сантиметра на два.", Recipe = recipeEntity };
-            var tip2 = new Tip() { Name = "Рис не перемешивать - аккуратно разровнять.", Recipe = recipeEntity };
-
-            _ingredientService.InsertIngredient(ingredient1);
-            _ingredientService.InsertIngredient(ingredient2);
-            _ingredientService.InsertIngredient(ingredient3);
-            _ingredientService.InsertIngredient(ingredient4);
-            _ingredientService.InsertIngredient(ingredient5);
-            _ingredientService.InsertIngredient(ingredient6);
-            _ingredientService.InsertIngredient(ingredient7);
-            _ingredientService.InsertIngredient(ingredient8);
-            _methodService.InsertMethod(method1);
-            _methodService.InsertMethod(method2);
-            _methodService.InsertMethod(method3);
-            _methodService.InsertMethod(method4);
-            _methodService.InsertMethod(method5);
-            _methodService.InsertMethod(method6);
-            _methodService.InsertMethod(method7);
-            _methodService.InsertMethod(method8);
-            _methodService.InsertMethod(method9);
-            _tipService.InsertTip(tip1);
-            _tipService.InsertTip(tip2);
-            var postEntity = new Post()
-            {
-                RecipeId = recipeEntity.Id
-            };
-            _postService.InsertPost(postEntity, userId);
-
-            return RedirectToAction("ShowMy");
+            ViewData["returnAction"] = "/Recipe/Show";
+            return View();
         }
     }
 }
